Current version is 0.16.279: Iris was updated.

+++ iris/src/irisnet/noncore/cutestuff/bsocket.h
@@ -53,7 +53,7 @@ public:
 	/*! Connect to a host via the specified protocol, or the default protocols if not specified */
 	void connectToHost(const QString &host, quint16 port, QAbstractSocket::NetworkLayerProtocol protocol = QAbstractSocket::UnknownNetworkLayerProtocol);
 	/*! Connect to the hosts for the specified service */
-	void connectToHost(const QString &service, const QString &transport, const QString &domain, quint16 port = std::numeric_limits<int>::max());
+	void connectToHost(const QString &service, const QString &transport, const QString &domain, quint16 port = std::numeric_limits<quint16>::max());
 	virtual QAbstractSocket* abstractSocket() const;
 	int socket() const;
 	void setSocket(int);
